Ajuste de variável divisão

+++ src/br/com/dio/calculadora/Calculadora.java
@@ -19,14 +19,13 @@ public class Calculadora {
 		int soma = soma(a,b);
 		int subtracao = subtracao(a,b);
 		int multiplicacao = multiplicacao(a,b);
-		int divisao = divisao(a,b);
+		double divisao = divisao(a,b);
 		
 		System.out.println("A soma:" + soma);
 		System.out.println("A Subtração:" + subtracao);
 		System.out.println("A Multiplicação:" + multiplicacao);
 		System.out.println("A Divisão:" + divisao);
-		
-		
+			
 	}
 	public static int soma(int a,int b) {
 		return a + b;
@@ -34,7 +33,7 @@ public class Calculadora {
 	public static int subtracao(int a,int b) {
 		return a - b;
 	}
-	public static int divisao(int a,int b) {
+	public static double divisao(double a,double b) {
 		return a / b;
 	}
 	public static int multiplicacao(int a,int b) {
